RavenDB-12477 : Handle revisions of documents that change their collection

+++ src/Raven.Server/Documents/Revisions/RevisionsStorage.cs
@@ -3,12 +3,10 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Text;
 using Raven.Client;
 using Raven.Client.Documents.Operations.Revisions;
-using Raven.Client.Exceptions.Documents;
-using Raven.Client.Extensions;
 using Raven.Client.ServerWide;
+using Raven.Server.Json;
 using Raven.Server.NotificationCenter.Notifications;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.Utils;
@@ -573,30 +571,59 @@ namespace Raven.Server.Documents.Revisions
             long numberOfRevisionsToDelete, TimeSpan? minimumTimeToKeep, string changeVector, long lastModifiedTicks)
         {
             long maxEtagDeleted = 0;
+            Table writeTable = null;
+            string currentCollection = null;
+            var deletedRevisionsCount = 0;
 
-            var deletedRevisionsCount = table.DeleteForwardFrom(RevisionsSchema.Indexes[IdAndEtagSlice], prefixSlice, true,
-                numberOfRevisionsToDelete,
-                deleted =>
+            while (true)
+            {
+                var hasValue = false;
+
+                foreach (var read in table.SeekForwardFrom(RevisionsSchema.Indexes[IdAndEtagSlice], prefixSlice, skip: 0, startsWith: true))
                 {
-                    var revision = TableValueToRevision(context, ref deleted.Reader);
+                    if (numberOfRevisionsToDelete <= deletedRevisionsCount)
+                        break;
 
-                    if (minimumTimeToKeep.HasValue &&
-                        _database.Time.GetUtcNow() - revision.LastModified <= minimumTimeToKeep.Value)
-                        return false;
-
-                    using (TableValueToSlice(context, (int)RevisionsTable.ChangeVector, ref deleted.Reader, out Slice key))
+                    using (TableValueReaderUtil.CloneTableValueReader(context, read.Result))
                     {
-                        var revisionEtag = TableValueToEtag((int)RevisionsTable.Etag, ref deleted.Reader);
-                        CreateTombstone(context, key, revisionEtag, collectionName, changeVector, lastModifiedTicks);
+                        var tvr = read.Result.Reader;
+                        var revision = TableValueToRevision(context, ref tvr);
+
+                        if (minimumTimeToKeep.HasValue &&
+                            _database.Time.GetUtcNow() - revision.LastModified <= minimumTimeToKeep.Value)
+                            return deletedRevisionsCount;
+
+                        hasValue = true;
+
+                        using (Slice.From(context.Allocator, revision.ChangeVector, out var keySlice))
+                        {
+                            CreateTombstone(context, keySlice, revision.Etag, collectionName, changeVector, lastModifiedTicks);
+
+                            maxEtagDeleted = Math.Max(maxEtagDeleted, revision.Etag);
+                            if ((revision.Flags & DocumentFlags.HasAttachments) == DocumentFlags.HasAttachments)
+                            {
+                                _documentsStorage.AttachmentsStorage.DeleteRevisionAttachments(context, revision, changeVector, lastModifiedTicks);
+                            }
+
+                            var docCollection = CollectionName.GetCollectionName(revision.Data);
+                            if (writeTable == null || docCollection != currentCollection)
+                            {
+                                currentCollection = docCollection;
+                                writeTable = EnsureRevisionTableCreated(context.Transaction.InnerTransaction, new CollectionName(docCollection));
+                            }
+
+                            writeTable.DeleteByKey(keySlice);
+                        }
                     }
 
-                    maxEtagDeleted = Math.Max(maxEtagDeleted, revision.Etag);
-                    if ((revision.Flags & DocumentFlags.HasAttachments) == DocumentFlags.HasAttachments)
-                    {
-                        _documentsStorage.AttachmentsStorage.DeleteRevisionAttachments(context, revision, changeVector, lastModifiedTicks);
-                    }
-                    return true;
-                });
+                    deletedRevisionsCount++;
+                    break;
+                }
+
+                if (hasValue == false)
+                    break;
+            }
+            
             _database.DocumentsStorage.EnsureLastEtagIsPersisted(context, maxEtagDeleted);
             return deletedRevisionsCount;
         }
